Add obstacle checker for pacman

+++ src/main/java/pacman/constant/MapConfig.java
@@ -1,10 +1,10 @@
 package pacman.constant;
 
 public class MapConfig {
-  public static final int WIDTH = 1250; // 25 grids in x
-  public static final int HEIGHT = 550; // 11 grids in y
+  public static final double WIDTH = 1250; // 25 grids in x
+  public static final double HEIGHT = 550; // 11 grids in y
 
-  public static final int GRID_LENGTH = 50;
+  public static final double GRID_LENGTH = 50;
 
-  public static final int STEP = 5;
+  public static final double STEP = 5;
 }
